add topics list and show api

+++ app/http/controllers/api/v1/topic_controller.go
@@ -15,8 +15,15 @@ type TopicsController struct {
 }
 
 func (ctrl *TopicsController) Index(c *gin.Context) {
-	topics := topic.All()
-	response.Data(c, topics)
+	request := requests.PaginationRequest{}
+	if ok := requests.Validate(c, &request, requests.Pagination); !ok {
+		return
+	}
+	data, pager := topic.Paginate(c, 10)
+	response.JSON(c, gin.H{
+		"data":  data,
+		"pager": pager,
+	})
 }
 
 func (ctrl *TopicsController) Show(c *gin.Context) {
@@ -85,10 +92,10 @@ func (ctrl *TopicsController) Delete(c *gin.Context) {
 		return
 	}
 
-	//if ok := policies.CanModifyTopic(c, topicModel); !ok {
-	//	response.Abort403(c)
-	//	return
-	//}
+	if ok := policies.CanModifyTopic(c, topicModel); !ok {
+		response.Abort403(c)
+		return
+	}
 
 	rowsAffected := topicModel.Delete()
 	if rowsAffected > 0 {
